requirements.txt
pytest==3.0.5
click==6.7
pathlib2==2.2.0
appdirs==1.4.2
GitPython==2.0.6
ruamel.yaml==0.13.7
libtmux==0.6.3
furl==0.5.6
utils2
google-cloud==0.23.0
pycountry==1.17
pyaml==16.12.2
splinter==0.7.5
six==1.10.0
